+++ src/events/handlers/device/common/alerts/lwtStatusAlertHandler.spec.js
@@ -7,7 +7,6 @@ chai.should();
 const expect = chai.expect;
 
 describe('lwt status alert handler', () => {
-
   let subject;
   let alertProvider;
   let socket;
@@ -17,28 +16,76 @@ describe('lwt status alert handler', () => {
     loggerStub = sinon.stub(logger, 'log');
 
     alertProvider = {
-      add: () => {},
+      update: () => {},
+      getLastAlertByKey: () => {},
     };
     socket = {};
     socket.emit = sinon.spy();
     subject = new LwtStatusAlertHandler(alertProvider, socket);
   });
 
-  it('should create alert for device online/offline', () => {
-    const addAlertStub = sinon.stub(alertProvider, 'add').returns(Promise.resolve());
+  afterEach(() => {
+    alertProvider.getLastAlertByKey.restore();
+    alertProvider.update.restore();
+  });
+
+  it('should create alert for device status if not present', () => {
+    sinon.stub(alertProvider, 'getLastAlertByKey').returns(Promise.resolve(null));
+    const updateAlertStub = sinon.stub(alertProvider, 'update').returns(Promise.resolve());
+
     const event = {
-      status: 'Online',
+      status: 'Offline',
       device: {
         gateway: 'gw1',
         deviceId: '22:44:77:33',
         description: 'device test',
-        status: { online: true },
+        status: { online: false },
       },
     };
 
     subject.process(event).then(() => {
-      addAlertStub.called.should.be.true;
-      addAlertStub.calledWith(sinon.match({ gateway: 'gw1', level: 'info', message: 'device test is ONLINE', deviceId: event.device.deviceId, type: 'Device offline' })).should.be.true;
+      updateAlertStub.calledWith(sinon.match({ gateway: 'gw1', level: 'info', message: 'device test is OFFLINE', deviceId: event.device.deviceId, type: 'Device_status' })).should.be.true;
+      socket.emit.called.should.be.true;
+    });
+  });
+
+  it('should update alert for device status if present', () => {
+    sinon.stub(alertProvider, 'getLastAlertByKey').returns(Promise.resolve({
+      gateway: 'gw1',
+      deviceId: '22:44:77:33',
+      message: 'device test is OFFLINE',
+      type: 'Device_status',
+      key: 'alert:Device_status:gw1:22:44:77:33',
+      date: new Date(),
+      read: false,
+      open: true,
+      level: 'info',
+    }));
+
+    const updateAlertStub = sinon.stub(alertProvider, 'update').returns(Promise.resolve());
+
+    const event = {
+      status: 'Offline',
+      device: {
+        gateway: 'gw1',
+        deviceId: '22:44:77:33',
+        description: 'device test',
+        status: { online: false },
+      },
+    };
+
+    subject.process(event).then(() => {
+      updateAlertStub.calledWith(sinon.match({
+        gateway: 'gw1',
+        deviceId: '22:44:77:33',
+        message: 'device test is OFFLINE',
+        type: 'Device_status',
+        key: 'alert:Device_status:gw1:22:44:77:33',
+        date: sinon.match.date,
+        read: false,
+        open: true,
+        lastUpdate: sinon.match.date,
+      })).should.be.true;
       socket.emit.called.should.be.true;
     });
   });
@@ -46,5 +93,4 @@ describe('lwt status alert handler', () => {
   after(() => {
     loggerStub.restore();
   });
-
 });